reseller

+++ app/controllers/reseller/users_controller.rb
@@ -1,7 +1,7 @@
 
 class Reseller::UsersController < Reseller::BaseController
   def index
-    @users = User.all
+    @users = User.where(parent_id: current_user.id)
   end
 
   def show
@@ -22,15 +22,46 @@ class Reseller::UsersController < Reseller::BaseController
   end
 
   def create
-    @user = User.new(:username => user_params[:username], :email => user_params[:email], :sms_count => user_params[:sms_count], :role => user_params[:role], :password =>'123456789', :password_confirmation => '123456789')
+    @user = User.new(:username => user_params[:username], :email => user_params[:email], :sms_count => user_params[:sms_count], :role => 0, :password =>user_params[:password], :password_confirmation => user_params[:password], :jasmin_password => user_params[:password], :parent_id => current_user.id)
     jasmin_user = JasminUser.new()
-    unless jasmin_user.add_user(user_params[:username], user_params[:username], user_params[:sms_count])
+    unless jasmin_user.add_user(user_params[:username], user_params[:password], user_params[:sms_count])
       render 'new'
     end
 
     if @user.save
+      @senders = params[:senders].split("\r\n")
+
+      @senders.each do |name|
+        @sender = Sender.new
+        @sender.name =  name
+        @sender.user_id = @user.id
+        if !@sender.save
+          render 'new'
+        end
+      end
+
+      jasmin_filter = JasminFilter.new()
+      unless jasmin_filter.add_filter(@user.id, "UserFilter", @user.username)
+        jasmin_delete_user = JasminUser.new()
+        jasmin_delete_user.delete_user(user_params[:username])
+        @user.delete
+        render 'new'
+        return
+      end
+
+      @filter = Filter.new(:fid => @user.id, :filter_type => "UserFilter", :parameter => @user.username)
+      unless @filter.save
+        jasmin_delete_user = JasminUser.new()
+        jasmin_delete_user.delete_user(user_params[:username])
+        jasmin_delete_filter = JasminFilter.new()
+        jasmin_delete_filter.delete_filter(@user.id)
+        @user.delete
+        render 'new'
+      end
       redirect_to reseller_users_path
     else
+      jasmin_delete_user = JasminUser.new()
+      jasmin_delete_user.delete_user(user_params[:username])
       render 'new'
     end
   end
@@ -38,10 +69,21 @@ class Reseller::UsersController < Reseller::BaseController
   def update
     @user = User.find(params[:id])
     jasmin_user = JasminUser.new()
-    unless jasmin_user.update_user(@user.username, user_params[:sms_count])
+    unless jasmin_user.update_user(@user.username, user_params[:jasmin_password], user_params[:sms_count])
       render 'new'
     end
-    if @user.update_attributes(user_params)
+    if @user.update_attributes(:email => user_params[:email], :sms_count => user_params[:sms_count], :role => 0, :password =>user_params[:jasmin_password], :password_confirmation => user_params[:jasmin_password], :jasmin_password => user_params[:jasmin_password])
+      Sender.where(:user_id => @user.id).delete_all
+      @senders = params[:senders].split("\r\n")
+
+      @senders.each do |name|
+        @sender = Sender.new
+        @sender.name =  name
+        @sender.user_id = @user.id
+        if !@sender.save
+          render 'new'
+        end
+      end
       redirect_to reseller_users_path, :notice => "User updated."
     else
       redirect_to reseller_users_path, :alert => "Unable to update user."
@@ -50,6 +92,15 @@ class Reseller::UsersController < Reseller::BaseController
 
   def destroy
     @user = User.find(params[:id])
+
+    jasmin_filter = JasminFilter.new()
+    unless jasmin_filter.delete_filter(@user.id)
+      redirect_to reseller_users_path, :notice => "Unable to delete user."
+      return
+    end
+    @filter = Filter.where(fid: @user.id)
+    @filter.destroy_all
+
     jasmin_user = JasminUser.new()
     unless jasmin_user.delete_user(@user.username)
       redirect_to reseller_users_path, :notice => "Unable to delete user."
@@ -67,6 +118,8 @@ class Reseller::UsersController < Reseller::BaseController
         :email,
         :sms_count,
         :role,
+        :password,
+        :jasmin_password,
     )
   end
 end
